test: move some of these test helpers to new homes without changes
