+++ test-integ/peering_commontopo/ac7_1_rotate_gw_test.go
@@ -8,10 +8,11 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/hashicorp/consul/api"
 	"github.com/hashicorp/consul/testing/deployer/topology"
 	"github.com/stretchr/testify/require"
 
-	"github.com/hashicorp/consul/api"
+	"github.com/hashicorp/consul/test-integ/topoutil"
 )
 
 // TestRotateGW ensures that peered services continue to be able to talk to their
@@ -145,7 +146,7 @@ func (s *suiteRotateGW) setup(t *testing.T, ct *commonTopo) {
 	if clu.Datacenter == agentlessDC {
 		nodeKind = topology.NodeKindDataplane
 	}
-	clu.Nodes = append(clu.Nodes, newTopologyMeshGatewaySet(
+	clu.Nodes = append(clu.Nodes, topoutil.NewTopologyMeshGatewaySet(
 		nodeKind,
 		"default",
 		s.newMGWNodeName,
